Split out fixture tests
